+++ crypto-onramp-example/src/main/java/com/stripe/android/crypto/onramp/example/OnrampTestTags.kt
@@ -6,6 +6,7 @@ internal const val LOGIN_LOGIN_BUTTON_TAG = "LoginLoginButtonTag"
 internal const val LOGIN_REGISTER_BUTTON_TAG = "LoginRegisterButtonTag"
 internal const val AUTHENTICATE_BUTTON_TAG = "AuthenticateButtonTag"
 internal const val REGISTER_WALLET_BUTTON_TAG = "RegisterWalletButtonTag"
+internal const val DELETE_WALLET_BUTTON_TAG_PREFIX = "DeleteWalletButtonTag-"
 internal const val COLLECT_CARD_BUTTON_TAG = "CollectCardButtonTag"
 internal const val CREATE_CRYPTO_TOKEN_BUTTON_TAG = "CreateCryptoTokenButtonTag"
 internal const val CREATE_SESSION_BUTTON_TAG = "CreateSessionButtonTag"
@@ -15,3 +16,7 @@ internal const val INITIALIZATION_FAILURE_ALERT_TITLE_TAG = "InitializationFailu
 internal const val INITIALIZATION_FAILURE_ALERT_MESSAGE_TAG = "InitializationFailureAlertMessageTag"
 internal const val SNACKBAR_TAG = "SnackbarTag"
 internal const val SNACKBAR_TEXT_TAG = "SnackbarTextTag"
+
+internal fun deleteWalletButtonTag(walletId: String): String {
+    return "$DELETE_WALLET_BUTTON_TAG_PREFIX$walletId"
+}