+++ crypto-onramp-example/src/main/java/com/stripe/android/crypto/onramp/example/ui/OnrampScreen.kt
@@ -91,6 +91,8 @@ internal fun OnrampScreen(
                     uiState = uiState,
                     onAuthenticate = onAuthenticateUser,
                     onRegisterWalletAddress = onRegisterWalletAddress,
+                    onDeleteWallet = viewModel::deleteWallet,
+                    onRefreshWallets = viewModel::refreshWallets,
                     onCollectKyc = viewModel::collectKycInfo,
                     onVerifyKyc = onVerifyKyc,
                     onStartVerification = onStartVerification,